requirements.txt
twilio
flask
python-dotenv
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib